api/requirements.txt
Django==1.11.3
djangorestframework==3.8.2
factory-boy==2.11.1